GuzzleClientAdapter moved to Utility namespace and implements HttpClientInterface
